// src/app/shared/models/model.ts
export interface counter {
  id: number;
  class: string;
  counter: number;
  title: string;
  end: string;
}
export interface favourite {
  id: number;
  title: string;
}
export interface tab1 {
  courseimg: string;
  avatar: string;
  title: string;
  content: string;
  students: string;
  currentamt: string;
}
export interface tab2 {
  courseimg: string;
  avatar: string;
  title: string;
  content: string;
  students: string;
  currentamt: string;
}
export interface tab3 {
  courseimg: string;
  avatar: string;
  title: string;
  content: string;
  students: string;
  currentamt: string;
}
export interface tab4 {
  courseimg: string;
  avatar: string;
  title: string;
  content: string;
  students: string;
  currentamt: string;
}
export interface tab5 {
  courseimg: string;
  avatar: string;
  title: string;
  content: string;
  students: string;
  currentamt: string;
}
export interface tab6 {
  courseimg: string;
  avatar: string;
  title: string;
  content: string;
  students: string;
  currentamt: string;
}
export interface tab7 {
  courseimg: string;
  avatar: string;
  title: string;
  content: string;
  students: string;
  currentamt: string;
}
export interface trending_course {
  badge: string;
  title: string;
  students: string;
  time: string;
  rate: string;
  avator: string;
  class: string;
}
export interface swiper {
  rating: string;
  avatar: string;
  name: string;
  role: string;
}